feat: set up links to game settings and quick start games
