--- src/css/components/_category-heading.scss
@@ -1,20 +1,0 @@
-.cmp-category-heading {
-	font-stretch: 125%;
-	font-size: clamp(1.5rem, 1.5rem + 0.5vw, 2.5rem);
-	font-weight: 280;
-	position: relative;
-	margin: 0;
-
-
-	&__link {
-		text-decoration: underline;
-		text-decoration-thickness: 1px;
-		text-underline-offset: 2px;
-		text-decoration-color: hsl(var(--smokey-black-hsl) / 0%);
-		transition: text-decoration-color 150ms ease-in-out;
-
-		&:hover {
-			text-decoration-color: hsl(var(--smokey-black-hsl) / 50%);
-		}
-	}
-}
