+++ src/css/styles.scss
@@ -63,7 +63,6 @@
 //    for example media object known from OOCSS.
 @forward 'objects/page';
 @forward 'objects/width-limiter';
-@forward 'objects/game-type';
 
 // =====================================================================
 // 6. Components
@@ -76,7 +75,6 @@
 @forward 'components/question';
 @forward 'components/answer';
 @forward 'components/categories';
-@forward 'components/category-heading';
 @forward 'components/multiple-choice';
 @forward 'components/game-header';
 @forward 'components/explanation-section';
