--- src/macros/categories.njk
@@ -1,11 +1,0 @@
-{% macro categories(catdir, title, questionData) %}
-<h3 id="{{ catdir }}" class="cmp-category-heading">
-	<a href="#{{ catdir }}" class="cmp-category-heading__link">{{ title }}</a>
-</h3>
-
-<ul class="cmp-categories">
-{% for category in questionData %}
-	{% include 'partials/category.njk' %}
-{% endfor %}
-</ul>
-{% endmacro %}
